+++ xgen_harness/orchestrator/__init__.py
@@ -1,5 +1,7 @@
 from .dag import DAGOrchestrator, AgentNode, DAGEdge, DAGResult
 from .multi_agent import MultiAgentExecutor
+from .complexity import ComplexityDetector, ComplexityVerdict
+from .multi_agent_planner import MultiAgentPlannerStage
 
 __all__ = [
     "DAGOrchestrator",
@@ -7,4 +9,7 @@ __all__ = [
     "DAGEdge",
     "DAGResult",
     "MultiAgentExecutor",
+    "ComplexityDetector",
+    "ComplexityVerdict",
+    "MultiAgentPlannerStage",
 ]